Added headers to source code files

+++ src/cpu/muj7cpu.h
@@ -1,3 +1,16 @@
+/**
+ * @file muj7cpu.h
+ *
+ * @brief Definition of muj7cpu structures and functions 
+ *
+ * Emulates the behaviour of muj7 CPU
+ *  3 general-purpose 8-bit registers (A, B, C)
+ *  flag 8-bit register (F)
+ *  program counter 16-bit register (PC)
+ *  19 instructions
+ *
+ * @see LICENSE for licensing details.
+ */
 
 #ifndef MUJ7_CPU_H
 #define MUJ7_CPU_H
@@ -8,7 +21,7 @@
 struct RegisterBank
 {
     uint8_t a, b, c, f;
-    uint8_t pc, sp;
+    uint16_t pc, sp;
 };
 
 struct CPU
